+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # OldAS ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added support for loading up the alternates for an article.
+  ([#2](https://github.com/davep/oldas/pull/2))
+
 ## v0.1.0
 
 **Released: 2026-01-01**